safari-600.7-branch/LayoutTests/sputnik/Conformance/12_Statement/12.6_Iteration_Statements/12.6.1_The_do_while_Statement/S12.6.1_A12-expected.txt

CONSOLE MESSAGE: line 78: SyntaxError: Unexpected keyword 'var'. Expected 'while' to end a do-while loop.
S12.6.1_A12

PASS Expected parsing failure

TEST COMPLETE
